+++ CHANGES.rst
@@ -1,8 +1,10 @@
-0.0.2 (unreleased)
+0.0.3 (2020-09-01)
 =====
+- Reimplemented to use jax.numpy.
+- Added testing for mesmer.seds and mesmer.likelihood.
+- Added example notebook implementing ML and HMC.
 
-
-0.0.2
+0.0.2 (2020-07-31)
 =====
 
 - Moved to CI / CD with GitHub Actions.